requirements.txt
python-hcl2==1.0.0
